# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.InfoResp:
    properties:
      build_time:
        type: string
      git_hash:
        type: string
      program_versions:
        $ref: "#/definitions/admin.ProgramVersions"
      uptime:
        type: string
      version:
        type: string
    type: object
  admin.ProgramVersions:
    properties:
      chat_downloader:
        type: string
      ffmpeg:
        type: string
      streamlink:
        type: string
      twitch_downloader:
        type: string
    type: object
...
  ent.User:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      oauth:
        description: Oauth holds the value of the "oauth" field.
        type: boolean
      role:
        allOf:
          - $ref: "#/definitions/utils.Role"
        description: Role holds the value of the "role" field.
      sub:
        description: Sub holds the value of the "sub" field.
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      username:
        description: Username holds the value of the "username" field.
        type: string
      webhook:
        description: Webhook holds the value of the "webhook" field.
        type: string
    type: object
...
  http.ChangePasswordRequest:
    properties:
      confirm_new_password:
        type: string
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
      - confirm_new_password
      - new_password
      - old_password
    type: object
...
  http.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
      - password
      - username
    type: object
  http.RegisterRequest:
    properties:
      password:
        minLength: 8
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
      - password
      - username
    type: object
...
  utils.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  utils.Role:
    enum:
      - admin
      - editor
      - archiver
      - user
    type: string
    x-enum-varnames:
      - AdminRole
      - EditorRole
      - ArchiverRole
      - UserRole
...
host: localhost:4000
info:
  contact:
    name: Zibbp
    url: https://github.com/zibbp/ganymede
  description: |-
    Authentication is handled using JWT tokens. The tokens are set as access-token and refresh-token cookies.
    For information regarding which role is authorized for which endpoint, see the http handler https://github.com/Zibbp/ganymede/blob/main/internal/transport/http/handler.go.
  license:
    name: GPL-3.0
  title: StreamVault API
  version: "1.0"
paths:
  /admin/info:
    get:
      consumes:
        - application/json
      description: Get ganymede info
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/admin.InfoResp"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      security:
        - ApiKeyCookieAuth: []
      summary: Get ganymede info
      tags:
        - admin
...
  /auth/change-password:
    post:
      consumes:
        - application/json
      description: Change password
      parameters:
        - description: Change password
          in: body
          name: change-password
          required: true
          schema:
            $ref: "#/definitions/http.ChangePasswordRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      security:
        - ApiKeyCookieAuth: []
      summary: Change password
      tags:
        - auth
  /auth/login:
    post:
      consumes:
        - application/json
      description:
        Login a user (sets access-token and refresh-token cookies). Access
        token lasts for 1 hour. Refresh token lasts for 1 month.
      parameters:
        - description: Login
          in: body
          name: login
          required: true
          schema:
            $ref: "#/definitions/http.LoginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/ent.User"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Login a user
      tags:
        - auth
  /auth/me:
    get:
      consumes:
        - application/json
      description: Get current user
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/ent.User"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      security:
        - ApiKeyCookieAuth: []
      summary: Get current user
      tags:
        - auth
  /auth/oauth/callback:
    get:
      consumes:
        - application/json
      description: OAuth callback for OAuth provider
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: OAuth callback
      tags:
        - auth
  /auth/oauth/login:
    get:
      consumes:
        - application/json
      description: Login a user with OAuth (sets access-token and refresh-token cookies)
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/ent.User"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Login a user with OAuth
      tags:
        - auth
...
  /auth/register:
    post:
      consumes:
        - application/json
      description: Register a user (does not log in)
      parameters:
        - description: Register
          in: body
          name: register
          required: true
          schema:
            $ref: "#/definitions/http.RegisterRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/ent.User"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Register a user
      tags:
        - auth
...
securityDefinitions:
  ApiKeyCookieAuth:
    in: cookie
    name: access-token
    type: apiKey
swagger: "2.0"