# ent/schema/user.go
package schema

import (
	"entgo.io/ent"
	"entgo.io/ent/schema/field"
	"github.com/google/uuid"
	"github.com/zibbp/ganymede/internal/utils"
	"time"
)

// User holds the schema definition for the User entity.
type User struct {
	ent.Schema
}

// Fields of the User.
func (User) Fields() []ent.Field {
	return []ent.Field{
		field.UUID("id", uuid.UUID{}).Default(uuid.New),
		field.String("sub").Unique().Optional(),
		field.String("username").Unique(),
		field.String("password").Sensitive().Optional(),
		field.Bool("oauth").Default(false),
		field.Enum("role").GoType(utils.Role("")).Default(string(utils.UserRole)),
		field.String("webhook").Optional(),
		field.String("api_key").Unique().Optional().Sensitive().Nillable(),
		field.Time("updated_at").Default(time.Now).UpdateDefault(time.Now),
		field.Time("created_at").Default(time.Now).Immutable(),
	}
}

// Edges of the User.
func (User) Edges() []ent.Edge {
	return nil
}